OGC Services are deprecated
I talked to Aaron about this and OGC Services are not supported in the traditional way anymore and should not be included in this table. There is a workaround to bring them back if needed but we'd rather motivate the users to use the new approach to create OGC conform web services and also advertise this in the training manual. Since both, WMS and WFS are published as data streaming service, 3.16 Other Self-Services would be a good place to mention the new OGC method. Here's a PR that shows how this was dealt with in the FME Server Admin Guide: http://bugzilla/show_bug.cgi?id=67079