Rearranged image
Based on feedback from 10 people at training - rearranged so that the image that shows how to do it is before the text. Many people asked  information regarding what was listted after the expression (ie, yuo dont need to type this, and the scrrenshot how it was done). So the solution should be moved some lines earlier as shown.